Move error struct to file
